+++ src/components/organisms/Editor/EditorMenu/ToolbarItem/ToolbarItem.tsx
@@ -1,3 +1,5 @@
+import { IconType } from 'react-icons';
+import { BsAlarm } from 'react-icons/bs';
 import { LabelButton } from './ToolbarButton/LabelButton';
 import { NormalButton } from './ToolbarButton/NormalButton';
 
@@ -5,24 +7,16 @@ export const ToolbarItem = (props: {
   name: string;
   type: string;
   buttonRef?: React.MutableRefObject<HTMLButtonElement | null>;
-  onClick?: (
-    e?: React.MouseEvent<HTMLButtonElement, MouseEvent>,
-    value?: string | number,
-  ) => void;
+  onClick?: (type: string, value?: string) => void;
   children?: React.ReactNode;
+  Icon?: IconType;
 }) => {
-  const {
-    name,
-    type,
-    buttonRef,
-    onClick,
-
-    children,
-  } = props;
+  const { name, type, buttonRef, onClick, Icon, children } = props;
   return type === 'LabelButton' ? (
     <LabelButton name={name} children={children} onClick={onClick} />
   ) : (
     <NormalButton
+      Icon={Icon || BsAlarm}
       name={name}
       children={children}
       buttonRef={buttonRef}
