+++ src/constants/editor.constant.ts
@@ -1,0 +1,33 @@
+export const FONT_SIZE_LIST: FontSizeList = {
+  type: 'FONT_SIZE',
+  list: [
+    {
+      value: '1',
+      fontSize: '10px',
+    },
+    {
+      value: '2',
+      fontSize: '13px',
+    },
+    {
+      value: '3',
+      fontSize: '16px',
+    },
+    {
+      value: '4',
+      fontSize: '18px',
+    },
+    {
+      value: '5',
+      fontSize: '24px',
+    },
+    {
+      value: '6',
+      fontSize: '32px',
+    },
+    {
+      value: '7',
+      fontSize: '48px',
+    },
+  ],
+};
